feat: Add logout functionality to profile screen
This commit adds a logout button to the profile screen, allowing users to sign out of the app. When the logout button is pressed, the user is signed out using the `FirebaseAuth` instance obtained from the dependency injection container. After signing out, the user is redirected to the sign-in screen.

Recent user commits:
- refactor: Initialize and dispose text controllers in SignUpPage
- chore: Update main.dart to use async main function
- add image splash screen
- chore: Add background color resource in colors.xml
- add image for splash screen
- chore: Update launch_background.xml to use custom splash screen image
- chore: Update launch_background.xml to use custom splash screen image
- chore: Update AndroidManifest.xml to use custom splash screen icon
- delete default image splash screen and icon app
- feat: Add profile screen with image picker for user profiles
- chore: Update homepage to handle auth state changes properly