feat: Add loading animation to profile screen image save
This commit adds a loading animation widget to the profile screen when saving the user's image. The loading animation provides visual feedback while the image is being uploaded to Firebase Storage and the user's profile information is being updated in Firestore. The loading animation widget is imported from the `loading_animation_widget` package.

Recent user commits:
- refactor: Initialize and dispose text controllers in SignUpPage
- chore: Update main.dart to use async main function
- add image splash screen
- chore: Add background color resource in colors.xml
- add image for splash screen
- chore: Update launch_background.xml to use custom splash screen image
- chore: Update launch_background.xml to use custom splash screen image
- chore: Update AndroidManifest.xml to use custom splash screen icon
- delete default image splash screen and icon app
- feat: Add profile screen with image picker for user profiles
- chore: Update homepage to handle auth state changes properly