fix: Update image source in launch_background.xml
This commit updates the image source in the `launch_background.xml` file to use the correct file name for the splash screen image. The previous file name was `splashscreen.png`, but it has been changed to `splashscreen`. This ensures that the correct image is displayed as the splash screen when the app is launched.

Recent user commits:
- change size
- feat: Add loading animation to profile screen image save
- refactor: Initialize and dispose text controllers in SignUpPage
- feat: Add loading animation to sign-up page
- feat: Add loading animation to sign-in page
- refactor: Update builder function in AppRouter
- chore: Update dependencies for cloud_firestore and loading_animation_widget
- feat: Add logout functionality to profile screen