feat: Add message sending functionality to ChatScreen
This commit adds the functionality to send messages in the ChatScreen. When the user enters a message and taps the send button, the message is sent to the Firestore database along with the user's information. The message is then displayed in the chat screen along with the sender's username and profile image.

Recent user commits:
- refactor: Initialize and dispose text controllers in SignUpPage
- refactor: Update FirebaseStorage registration in injection.dart
- fix: Update image source in launch_background.xml
- change size
- feat: Add loading animation to profile screen image save
- feat: Add loading animation to sign-up page
- feat: Add loading animation to sign-in page
- refactor: Update builder function in AppRouter
- chore: Update dependencies for cloud_firestore and loading_animation_widget
- feat: Add logout functionality to profile screen