feat: Add ChatView widget for displaying chat messages
This commit adds the ChatView widget, which is responsible for displaying the chat messages in the chat screen. The widget uses a StreamBuilder to listen for changes in the Firestore collection and displays the messages in a ListView. It also handles different connection states and error scenarios. The ChatView widget enhances the user experience by providing a real-time chat interface.

Recent user commits:
- refactor: Update CustomButtons to be a StatelessWidget
- feat: Add ChatBubble widget for displaying chat messages
- feat: Add message sending functionality to ChatScreen
- refactor: Initialize and dispose text controllers in SignUpPage
- refactor: Update FirebaseStorage registration in injection.dart
- fix: Update image source in launch_background.xml
- change size
- feat: Add loading animation to profile screen image save
- feat: Add loading animation to sign-up page
- feat: Add loading animation to sign-in page
- refactor: Update builder function in AppRouter
- chore: Update dependencies for cloud_firestore and loading_animation_widget
- feat: Add logout functionality to profile screen