feat: Add lazy loading for images in ChatView
This commit adds lazy loading functionality for images in the ChatView widget. By implementing lazy loading, the images will only be loaded when they are visible on the screen, improving the page load performance and reducing the amount of data transferred. This enhancement enhances the user experience by providing a smoother and faster chat interface.

Recent user commits:
- refactor: Update CustomButtons to be a StatelessWidget
- feat: Add ChatBubble widget for displaying chat messages
- feat: Add message sending functionality to ChatScreen
- refactor: Initialize and dispose text controllers in SignUpPage
- refactor: Update FirebaseStorage registration in injection.dart
- fix: Update image source in launch_background.xml
- change size
- feat: Add loading animation to profile screen image save
- feat: Add loading animation to sign-up page
- feat: Add loading animation to sign-in page
- refactor: Update builder function in AppRouter
- chore: Update dependencies for cloud_firestore and loading_animation_widget
- feat: Add logout functionality to profile screen